Use react-router-dom's BrowserRouter, Routes and Route instead of createBrowserRouter, RouterProvider, and Outlet. In order to not increase bundle size for calypso independent code parts which use react-router-dom
