--- client/landing/subscriptions/components/tab-view/index.ts
@@ -1,1 +1,0 @@
-export { default as TabView } from './tab-view';
