+++ client/landing/subscriptions/components/tab-views/comments/comments.tsx
@@ -1,4 +1,4 @@
-import { TabView } from '../../../../components/tab-view';
+import TabView from '../tab-view';
 
 const Comments = () => (
 	<TabView errorMessage="" isLoading={ false }>
