+++ client/landing/subscriptions/components/tab-views/index.ts
@@ -1,0 +1,3 @@
+export { Comments } from './comments';
+export { Settings } from './settings';
+export { Sites } from './sites';
