+++ client/landing/subscriptions/components/tabs-switcher/index.tsx
@@ -1,0 +1,1 @@
+export { default as TabsSwitcher } from './tabs-switcher';
