+++ client/landing/subscriptions/components/tabs-switcher/styles.scss
