--- client/landing/subscriptions/components/tabs/index.tsx
@@ -1,1 +1,0 @@
-export { default as Tabs } from './tabs';
