--- client/landing/subscriptions/router/index.ts
@@ -1,1 +1,0 @@
-export { default as createRouter } from './create-router';
